2024-11-19 | fixed SQL

+++ order-service/src/main/resources/sql/schema-mysql.sql
@@ -3,7 +3,7 @@ CREATE TABLE orders
     id              BIGINT AUTO_INCREMENT PRIMARY KEY,
     user_id         BIGINT NOT NULL,
     status          VARCHAR(255) NOT NULL,
-    total_amount    INTEGER,
+    total_amount    DECIMAL(10,2),
     created_at      TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     updated_at      TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP
 );
@@ -14,6 +14,6 @@ CREATE TABLE order_items
     order_id        BIGINT NOT NULL,
     product_id      BIGINT NOT NULL,
     quantity        INT NOT NULL,
-    price           INTEGER NOT NULL,
+    price           DECIMAL(10,2) NOT NULL,
     FOREIGN KEY     (order_id) REFERENCES orders(id)
 );
